Fix some bugs in todo queries, and be smarter and snoozing repeated tasks
In particular, we look at how far from the previous deadline (calculated as
current deadline - repeat interval) the snooze was. So if you had snoozed until
a week before the deadline, we will snooze until a week before the next
deadline. This should be a relatively lightweight way of fine-tuning recurring
tasks, so you can, for example, get a rent reminder happening only starting a
week before the first, but have a bill reminder start a few days into the month.